Fix usage of tracepoint API for Linux kernel 3.15.0 and newer
In Linux kernel version 3.15.0, the tracepoint API changed and the
preferred way is to use for_each_kernel_tracepoint() to iterate the
built-in tracepoints, and use the new tracepoint module notifier to
examine mod->tracepoints_ptrs[] for the rest.

Add new implementation for using tracepoints, sense_tracepoint.c,
for sense in the linux-module derived almost verbatim from
[lttng](https://lttng.org/).

Update Makefile and sense.c, in which tracepoints are used, to use the
new sense_tracepoint.h interface.